requirements.txt
streamlit==1.14.0
langchain==0.0.56
openai==0.11.0
requests==2.27.1
folium==0.12.1
streamlit-folium==0.8.1
geopy==2.2.0
python-dotenv==0.19.2
pandas==1.3.5
geopandas==0.9.0
shapely==1.7.1
